requirements.txt
pytest>=7.0.0
python-dotenv>=0.19.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
google-generativeai>=0.3.2
pydantic>=2.0.0